backend/requirements.txt
fastapi
pymongo
python-dotenv
openai
Pillow
python-multipart
sqlalchemy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

#MongoDB Cluster for Varaad,Rutuj
